+++ R/roxygen.R
@@ -50,6 +50,10 @@ NULL
 #'   Matrix of predicted probabilities, each column is a vector of probabilities for a
 #'   specific class label.
 #'   Columns must be named with levels of `truth`.
+#' @param sample_weights :: `numeric()`\cr
+#'   Non-negative sample weights.
+#'   Must have the same levels and length as `truth`.
+#'   Defaults to equal sample weights.
 #' @param na_value :: `numeric(1)`\cr
 #'   Value that should be returned if the measure is not defined for the input
 #'   (as described in the note). Default is `NaN`.