+++ tests/testthat/test_classif.R
@@ -70,6 +70,24 @@ test_that("tests from Metrics", {
   # expect_equal(kappa, 0.624536446425734)
 })
 
+test_that("bacc", {
+  truth = factor(c("a", "a", "b", "b"), levels = c("a", "b"))
+  response = factor(c("a", "a", "b", "a"), levels = c("a", "b"))
+  expect_equal(bacc(truth, response), 0.75)
+  expect_equal(bacc(truth, response, sample_weights = c(0.25, 0.25, 0.25, 0.25)), 0.75)
+  expect_equal(bacc(truth, response, sample_weights = c(0.25, 0.25, 0.25, 1)), 0.6)
+
+  truth = factor(c("a", "a", "a", "a", "a", "b"), levels = c("a", "b"))
+  response = factor(c("a", "a", "a", "a", "b", "b"), levels = c("a", "b"))
+  expect_equal(bacc(truth, response), 0.9)
+  expect_equal(bacc(truth, response, sample_weights = c(0, 0, 0, 0, 0, 1)), 1)
+  expect_equal(bacc(truth, response, sample_weights = c(0, 0, 0, 0, 0.5, 0.5)), 0.5)
+
+  truth = factor(c("c", "a", "a", "a", "a", "b"), levels = c("a", "b", "c"))
+  response = factor(c("c", "a", "a", "a", "b", "b"), levels = c("a", "b", "c"))
+  expect_equal(round(bacc(truth, response), 3), 0.917)
+})
+
 # test_that("ber", {
 #   truth = factor(c("a", "a", "b", "b", "c", "c"), levels = c("a", "b", "c"))
 #   response = factor(c("a", "a", "b", "b", "c", "c"), levels = c("a", "b", "c"))