+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed some tests which stochastically failed.
 * The name / title of the measure is now also stored in the meta data.
+* Added new measure `bacc` (Balanced Accuracy).
 
 # mlr3measures 0.1.0